+++ src/integrations/GoogleAds/browser.js
@@ -1,6 +1,5 @@
 /* eslint-disable class-methods-use-this */
 import logger from '../../utils/logUtil';
-import { LOAD_ORIGIN } from '../../utils/ScriptLoader';
 import {
   getHashFromArrayWithDuplicate,
   removeUndefinedAndNullValues,
@@ -12,6 +11,7 @@ import {
   getConversionData,
 } from './utils';
 import { NAME } from './constants';
+import { loader } from './loader';
 
 class GoogleAds {
   constructor(config, analytics, destinationInfo) {
@@ -46,18 +46,7 @@ class GoogleAds {
 
   init() {
     const sourceUrl = `https://www.googletagmanager.com/gtag/js?id=${this.conversionId}`;
-    (function (id, src, document) {
-      logger.debug(`in script loader=== ${id}`);
-      const js = document.createElement('script');
-      js.src = src;
-      js.async = 1;
-      js.setAttribute('data-loader', LOAD_ORIGIN);
-      js.type = 'text/javascript';
-      js.id = id;
-      const e = document.getElementsByTagName('head')[0];
-      logger.debug('==script==', e);
-      e.appendChild(js);
-    })('googleAds-integration', sourceUrl, document);
+    loader('googleAds-integration', sourceUrl, document);
 
     window.dataLayer = window.dataLayer || [];
     window.gtag = function () {
